Relocate react-modal from AddBook to Result (commented out), add conditional confirmation message to Result
